+++ src/main/java/com/example/water/domain/comment/repository/CommentRepository.java
@@ -1,9 +1,10 @@
 package com.example.water.domain.comment.repository;
 
-import com.example.water.domain.comment.Comment;
-import com.example.water.domain.user.User;
+import com.example.water.domain.comment.entity.Comment;
+import com.example.water.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
     Long countByUserId(User user);
+    Comment findFirstByUserIdOrderByDate(User user);
 }
