+++ src/main/java/com/example/water/domain/comment/service/CommentService.java
@@ -5,16 +5,16 @@ import com.example.water.domain.chatGpt.config.ChatGptConfig;
 import com.example.water.domain.chatGpt.dto.ChatGptQuestionRequest;
 import com.example.water.domain.chatGpt.dto.ChatGptRequest;
 import com.example.water.domain.chatGpt.dto.ChatGptResponse;
-import com.example.water.domain.comment.Comment;
+import com.example.water.domain.comment.entity.Comment;
 import com.example.water.domain.comment.dto.request.CommentRequest;
 import com.example.water.domain.comment.dto.request.SaveRequest;
 import com.example.water.domain.comment.repository.CommentRepository;
-import com.example.water.domain.crystal.Crystal;
+import com.example.water.domain.crystal.entity.Crystal;
 import com.example.water.domain.crystal.repository.CrystalRepository;
 import com.example.water.domain.emotion.Emotion;
 import com.example.water.domain.emotion.repository.EmotionRepository;
-import com.example.water.domain.user.User;
-import com.example.water.domain.user.UserRepository;
+import com.example.water.domain.user.entity.User;
+import com.example.water.domain.user.repository.UserRepository;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
