--- src/main/java/com/example/water/domain/crystal/CrystalRepository.java
@@ -1,9 +1,0 @@
-package com.example.water.domain.crystal;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CrystalRepository extends JpaRepository<Crystal, Long> {
-
-}
