+++ src/main/java/com/example/water/domain/crystal/controller/CrystalController.java
@@ -1,12 +1,12 @@
-package com.example.water.domain.crystal;
+package com.example.water.domain.crystal.controller;
 
-import com.example.water.domain.comment.CommentService;
-import com.example.water.domain.comment.DTO.CommentResponseDTO;
-import com.example.water.domain.crystal.DTO.CrystalResponseDTO;
+import com.example.water.domain.comment.service.CommentService;
+import com.example.water.domain.crystal.service.CrystalService;
+import com.example.water.domain.crystal.dto.response.CrystalResponse;
 import com.example.water.global.BaseResponse;
 import com.example.water.global.ErrorCode;
 import com.example.water.global.SuccessCode;
-import com.example.water.global.auth.KakaoService;
+import com.example.water.global.auth.service.KakaoService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -22,12 +22,12 @@ public class CrystalController {
     private final CommentService commentService;
 
     @GetMapping("/all")
-    public ResponseEntity<BaseResponse<List<CrystalResponseDTO>>> getAllCrystal(@RequestHeader("Authorization") String authorizationHeader) {
+    public ResponseEntity<BaseResponse<List<CrystalResponse>>> getAllCrystal(@RequestHeader("Authorization") String authorizationHeader) {
         try {
             String access_token=authorizationHeader.substring(7);
             Map<String, Object> userInfo = kakaoService.getUserInfo(access_token);
             Long userId= (Long) userInfo.get("user_id");
-            List<CrystalResponseDTO> crystalResponses = crystalService.getCrystalResponses(userId);
+            List<CrystalResponse> crystalResponses = crystalService.getCrystalResponses(userId);
 
             return ResponseEntity.ok(BaseResponse.success(SuccessCode.CUSTOM_SUCCESS, crystalResponses));
         } catch (Exception e) {
