+++ src/main/java/com/example/water/domain/crystal/dto/response/CrystalResponse.java
@@ -1,4 +1,4 @@
-package com.example.water.domain.crystal.DTO;
+package com.example.water.domain.crystal.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Getter
 @Builder
 @AllArgsConstructor
-public class CrystalResponseDTO {
+public class CrystalResponse {
     private Long crystalId;
     private Long red;
     private Long green;
