+++ src/main/java/com/example/water/domain/crystal/repository/CrystalRepository.java
@@ -1,6 +1,6 @@
 package com.example.water.domain.crystal.repository;
 
-import com.example.water.domain.crystal.Crystal;
+import com.example.water.domain.crystal.entity.Crystal;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CrystalRepository extends JpaRepository<Crystal, Long> {
