+++ src/main/java/com/example/water/domain/diary/service/DiaryService.java
@@ -3,8 +3,8 @@ package com.example.water.domain.diary.service;
 import com.example.water.domain.comment.repository.CommentRepository;
 import com.example.water.domain.diary.dto.response.DiaryResponse;
 import com.example.water.domain.diary.dto.request.DiaryRequest;
-import com.example.water.domain.user.User;
-import com.example.water.domain.user.UserRepository;
+import com.example.water.domain.user.entity.User;
+import com.example.water.domain.user.repository.UserRepository;
 import io.github.flashvayne.chatgpt.service.ChatgptService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
