+++ src/main/java/com/example/water/domain/user/dto/Mypage.java
@@ -1,4 +1,4 @@
-package com.example.water.domain.user.DTO;
+package com.example.water.domain.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Getter
 @Builder
 @AllArgsConstructor
-public class MypageDTO {
+public class Mypage {
     private Long crystalCount;
     private Long sinceDate;
     private String nickname;
