+++ src/main/java/com/example/water/domain/user/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.example.water.domain.user;
+package com.example.water.domain.user.repository;
 
+import com.example.water.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
